push truncate column  code
