--- docs/1.17.0/IMPLEMENTATION_SUMMARY.md
@@ -1,254 +1,0 @@
-# 结果集字段截取功能开发总结
-
-## 版本信息
-- **版本号**: 1.17.0
-- **开发日期**: 2025-10-28
-- **功能名称**: 结果集查看、下载和导出接口优化 - 超长字段截取功能
-
-## 一、需求概述
-
-### 背景
-当前结果集查看功能存在问题:当某一列字段内容超过10000字符时,会导致结果集无法正常查看。
-
-### 目标
-为结果集查看、下载、导出功能增加超长字段检测和截取能力,提升系统稳定性和用户体验。
-
-## 二、功能需求
-
-1. **结果集查看**: 超过10000字符时检测并截取
-2. **结果集下载**: 超过10000字符时检测并截取
-3. **结果集导出**: 超过32767字符时检测并截取
-4. **超长字段限制**: 最多收集20个超长字段信息
-5. **功能开关**: 所有功能可通过配置开关控制
-
-## 三、实现方案
-
-### 3.1 架构设计
-```
-FsRestfulApi (REST API Layer)
-    ↓
-FieldTruncationHelper (工具类)
-    ↓
-配置项 (WorkSpaceConfiguration)
-```
-
-### 3.2 核心组件
-
-#### 配置类
-**文件**: `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/conf/WorkSpaceConfiguration.java`
-
-新增配置项:
-- `FIELD_TRUNCATION_ENABLED`: 功能总开关,默认 `false`
-- `FIELD_VIEW_MAX_LENGTH`: 查看最大长度,默认 `10000`
-- `FIELD_DOWNLOAD_MAX_LENGTH`: 下载最大长度,默认 `10000`
-- `FIELD_EXPORT_MAX_LENGTH`: 导出最大长度,默认 `32767`
-- `OVERSIZED_FIELD_MAX_COUNT`: 最多收集数量,默认 `20`
-
-#### 实体类
-
-1. **OversizedFieldInfo**
-   - 位置: `org.apache.linkis.filesystem.entity.OversizedFieldInfo`
-   - 字段: fieldName, rowIndex, actualLength, maxLength
-
-2. **FieldTruncationResult**
-   - 位置: `org.apache.linkis.filesystem.entity.FieldTruncationResult`
-   - 字段: hasOversizedFields, oversizedFields, maxOversizedFieldCount, data
-
-#### 工具类
-
-**FieldTruncationHelper**
-- 位置: `org.apache.linkis.filesystem.util.FieldTruncationHelper`
-- 核心方法:
-  - `detectAndHandle()`: 检测并处理超长字段
-  - `detectOversizedFields()`: 检测超长字段
-  - `truncateFields()`: 截取超长字段
-
-#### API改造
-
-1. **openFile() 方法**
-   - 文件: `FsRestfulApi.java`
-   - 新增参数: `truncateOversizedFields` (Boolean, 可选)
-   - 功能: 在返回数据前检测和截取超长字段
-   - 返回扩展字段:
-     - `hasOversizedFields`: 是否存在超长字段
-     - `oversizedFields`: 超长字段列表
-     - `maxOversizedFieldCount`: 最大收集数量
-
-2. **resultsetToExcel() 方法**
-   - 文件: `FsRestfulApi.java`
-   - 新增参数: `truncateOversizedFields` (Boolean, 可选)
-   - 功能: 导出前检测和截取超长字段(32767字符)
-
-## 四、文件清单
-
-### 新建文件
-1. `docs/1.17.0/requirements/resultset-field-truncation.md` - 需求文档
-2. `docs/1.17.0/design/resultset-field-truncation-design.md` - 设计文档
-3. `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/entity/OversizedFieldInfo.java`
-4. `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/entity/FieldTruncationResult.java`
-5. `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/util/FieldTruncationHelper.java`
-
-### 修改文件
-1. `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/conf/WorkSpaceConfiguration.java`
-   - 新增5个配置项
-
-2. `linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/restful/api/FsRestfulApi.java`
-   - openFile() 方法: 新增参数和字段截取逻辑
-   - resultsetToExcel() 方法: 新增参数和字段截取逻辑
-   - 新增 import 语句
-
-## 五、配置说明
-
-### 启用功能
-在 linkis 配置文件中添加:
-```properties
-# 启用字段截取功能
-linkis.resultset.field.truncation.enabled=true
-
-# 查看功能字段最大长度
-linkis.resultset.field.view.max.length=10000
-
-# 下载功能字段最大长度
-linkis.resultset.field.download.max.length=10000
-
-# 导出功能字段最大长度(Excel限制)
-linkis.resultset.field.export.max.length=32767
-
-# 最多收集超长字段数量
-linkis.resultset.field.oversized.max.count=20
-```
-
-## 六、API使用说明
-
-### 6.1 结果集查看
-
-**请求示例1: 仅检测**
-```
-GET /api/rest_j/v1/filesystem/openFile?path=/path/to/result&page=1&pageSize=100
-```
-
-**返回示例**:
-```json
-{
-  "status": 0,
-  "data": {
-    "metadata": [...],
-    "fileContent": [...],
-    "hasOversizedFields": true,
-    "oversizedFields": [
-      {
-        "fieldName": "description",
-        "rowIndex": 5,
-        "actualLength": 15000,
-        "maxLength": 10000
-      }
-    ],
-    "maxOversizedFieldCount": 20
-  }
-}
-```
-
-**请求示例2: 执行截取**
-```
-GET /api/rest_j/v1/filesystem/openFile?path=/path/to/result&page=1&pageSize=100&truncateOversizedFields=true
-```
-
-### 6.2 结果集导出
-
-**请求示例**:
-```
-GET /api/rest_j/v1/filesystem/resultsetToExcel?path=/path/to/result&outputFileType=xlsx&truncateOversizedFields=true
-```
-
-## 七、前后端交互流程
-
-```
-1. 前端请求查看结果集(不传truncateOversizedFields参数)
-2. 后端检测到超长字段,返回hasOversizedFields=true和超长字段列表
-3. 前端展示提示弹窗:"发现5个超长字段,是否截取前10000个字符?"
-4. 用户确认后,前端重新请求(truncateOversizedFields=true)
-5. 后端执行截取并返回截取后的数据
-6. 前端正常展示结果集
-```
-
-## 八、遵循的开发规则
-
-✅ **最小改动原则**
-- 仅在API层添加逻辑,不修改底层存储代码
-- 不修改现有接口签名,仅添加可选参数
-
-✅ **功能可配置原则**
-- 所有功能通过 `FIELD_TRUNCATION_ENABLED` 开关控制
-- 开关关闭时,行为与原版本完全一致
-
-✅ **配置管理规则**
-- 使用 `CommonVars` 统一管理配置
-- 配置放在模块的 conf 目录下
-
-✅ **代码边界约束**
-- 未修改数据库结构
-- 未引入新的第三方依赖
-- 未修改现有公共接口签名
-
-✅ **无需前端开发**
-- 仅实现后端接口功能
-
-## 九、测试建议
-
-### 9.1 功能测试
-1. 功能开关关闭时,验证行为与原版本一致
-2. 功能开关开启时,测试检测功能
-3. 测试截取功能(查看、导出)
-4. 测试超过20个超长字段的场景
-5. 测试边界值(10000, 32767字符)
-
-### 9.2 性能测试
-1. 测试大结果集(1000行+)的检测性能
-2. 测试大结果集的截取性能
-3. 对比功能开启前后的响应时间
-
-### 9.3 兼容性测试
-1. 老版本前端访问新接口
-2. 不传新参数时的表现
-3. 传入新参数时的表现
-
-## 十、已知限制
-
-1. **下载功能**: 当前版本未实现下载功能的截取(保留为后续优化项)
-2. **导出元数据**: 导出功能的元数据写入需要进一步测试
-3. **性能影响**: 大结果集的全量检测可能影响性能,建议配合分页使用
-
-## 十一、后续优化方向
-
-1. 实现下载功能的字段截取
-2. 优化导出功能的元数据处理
-3. 添加只检测前N行的配置项(避免全量检测)
-4. 支持异步检测大结果集
-5. 前端实现完整的交互逻辑
-
-## 十二、注意事项
-
-⚠️ **重要提醒**
-1. 功能默认关闭,需要手动开启
-2. 开启后会影响结果集查看性能,建议先在测试环境验证
-3. 截取会丢失部分数据,用户需要明确知晓
-4. Excel导出的32767字符限制是Excel自身的限制
-
-## 十三、编译部署
-
-### 编译
-```bash
-cd linkis-public-enhancements/linkis-pes-publicservice
-mvn clean package -DskipTests
-```
-
-### 部署
-1. 替换 linkis-pes-publicservice jar包
-2. 重启 linkis-ps-publicservice 服务
-3. 修改配置文件启用功能
-4. 验证功能是否正常
-
----
-
-**开发者**: Claude Code
-**日期**: 2025-10-28
