+++ linkis-public-enhancements/linkis-pes-publicservice/src/main/java/org/apache/linkis/filesystem/conf/WorkSpaceConfiguration.java
@@ -83,18 +83,6 @@ public class WorkSpaceConfiguration {
   public static final CommonVars<Boolean> FIELD_TRUNCATION_ENABLED =
       CommonVars$.MODULE$.apply("linkis.resultset.field.truncation.enabled", false);
 
-  public static final CommonVars<Integer> FIELD_VIEW_MAX_LENGTH =
-      CommonVars$.MODULE$.apply("linkis.resultset.field.view.max.length", 10000);
-
-  public static final CommonVars<Integer> FIELD_DOWNLOAD_MAX_LENGTH =
-      CommonVars$.MODULE$.apply("linkis.resultset.field.download.max.length", 10000);
-
-  public static final CommonVars<Integer> FIELD_EXPORT_MAX_LENGTH =
-      CommonVars$.MODULE$.apply("linkis.resultset.field.export.max.length", 32767);
-
-  public static final CommonVars<Integer> OVERSIZED_FIELD_MAX_COUNT =
-      CommonVars$.MODULE$.apply("linkis.resultset.field.oversized.max.count", 20);
-
   public static final ExecutorService executorService =
       new ThreadPoolExecutor(
           FILESYSTEM_FS_THREAD_NUM.getValue(),
